httpd: Add highlevel parselib
