+++ src/cmds/net/httpd/httpd.c
@@ -35,33 +35,6 @@
 static char httpd_g_inbuf[BUFF_SZ];
 static char httpd_g_outbuf[BUFF_SZ];
 
-static int httpd_read_http_header(const struct client_info *cinfo, char *buf, size_t buf_sz) {
-	const int sk = cinfo->ci_sock;
-	const char *pattern = "\r\n\r\n";
-	char pattbuf[strlen("\r\n\r\n")];
-	char *pb;
-
-	pb = buf;
-	if (0 > read(sk, pattbuf, sizeof(pattbuf))) {
-		return -errno;
-	}
-	while (0 != strncmp(pattern, pattbuf, sizeof(pattbuf)) && buf_sz > 0) {
-		*(pb++) = pattbuf[0];
-		buf_sz--;
-		memmove(pattbuf, pattbuf + 1, sizeof(pattbuf) - 1);
-		if (0 > read(sk, &pattbuf[sizeof(pattbuf) - 1], 1)) {
-			return -errno;
-		}
-	}
-
-	if (buf_sz == 0) {
-		return -ENOENT;
-	}
-
-	memcpy(pb, pattbuf, sizeof(pattbuf));
-	return pb + sizeof(pattbuf) - buf;
-}
-
 static int httpd_header(const struct client_info *cinfo, int st, const char *msg) {
 	FILE *skf = fdopen(cinfo->ci_sock, "rw");
 
@@ -107,22 +80,13 @@ static void httpd_on_cgi_child(const struct client_info *cinfo, pid_t child) {
 	}
 }
 
-static void httpd_client_process(const struct client_info *cinfo) {
+static void httpd_client_process(struct client_info *cinfo) {
 	struct http_req hreq;
 	pid_t cgi_child;
-	int ret;
+	int err;
 
-	ret = httpd_read_http_header(cinfo, httpd_g_inbuf, sizeof(httpd_g_inbuf) - 1);
-	if (ret < 0) {
-		log_error("can't read from client socket: %s", strerror(errno));
-		return;
-	}
-	httpd_g_inbuf[ret] = '\0';
-
-	memset(&hreq, 0, sizeof(hreq));
-	if (NULL == httpd_parse_request(httpd_g_inbuf, &hreq)) {
-		log_error("can't parse request");
-		return;
+	if ((err = httpd_build_request(cinfo, &hreq, httpd_g_inbuf, sizeof(httpd_g_inbuf)))) {
+		log_error("can't build request: %s", strerror(-err));
 	}
 
 	log_debug("method=%s uri_target=%s uri_query=%s",
